Move tables to tables directory
