+++ views/tables/ParticipantsTable/index.jsx
@@ -3,8 +3,8 @@ import { Checkbox } from '@mui/material'
 import StarBorder from '@mui/icons-material/StarBorder'
 import Star from '@mui/icons-material/Star'
 
-import Table from '../../Table'
-import { SORT_DIRECTION } from '../../../../constants'
+import Table from '../Table'
+import { SORT_DIRECTION } from '../../../constants'
 
 const ParticipantsTable = (props) => {
   const {
